Partial fix for #158 .
When an element is compared it now first checks if the other element has the 'type' property, avoiding a MissingPropertyException. When a choice is compared it now first check if the other entity has the 'elements' propery, avoiding a MissingPropertyException.

A Difference is created on both occasions, but without a proper message.